+++ packages/agent-chat/src/views/Settings/DocSettings.vue
@@ -35,12 +35,6 @@ const securitySchemes = computed(
 <template>
   <div class="docSettings">
     <div>
-      <ServerSelector
-        :eventBus
-        :selectedServer
-        :servers="document.servers ?? []" />
-    </div>
-    <div>
       <Auth
         :document
         :environment
@@ -49,6 +43,12 @@ const securitySchemes = computed(
         :options="config"
         :securitySchemes
         :selectedServer />
+    </div>
+    <div>
+      <ServerSelector
+        :eventBus
+        :selectedServer
+        :servers="document.servers ?? []" />
     </div>
   </div>
 </template>